+++ pass_expand.go
@@ -53,6 +53,10 @@ func expandEntity(m *Munch, b Bindings, n *Node) ([]*Node, error) {
 			impl.List = append(impl.List, op)
 			continue
 		}
+		if op.IsList && contains(knownNeeds, op.Head()) { // a need, like (durable)
+			impl.List = append(impl.List, L(Sym("needs"), Sym(op.Head())))
+			continue
+		}
 		meth, err := storeMethod(op, name, fields)
 		if err != nil {
 			return nil, err